refactor xtask no-blob into srp modules
